Use the full document of the rST response. Backwards incompatible. Refs #14.
This changes how the document is fetched from the rST renderer. Before it was fetching the actual document fragment while now it fetches the entire document body including the main and subtitle.

If you prefer the previous behavior, this is easily changed by subclassing the Filter class and adjust the part name:

```
class OldStyleRstMarkupFilter(RstMarkupFilter):
    rst_part_name = 'fragment'
```